fix: Fix front-end startup stuck
When the disk manager starts, it will send an quit message to the
background service, then start a new background service, and wait for
the new background service to start before displaying the main
interface.

In the extreme case, when the old service has not exited and the new
service has not been started, the front end detects the existence of
the service, then displays the main interface and sends a request to
the old service. Since the old service may have exited during the
execution of the request, the front end cannot receive the
updateDeviceInfo signal and remains in the initialization interface.

The solution is that the front end needs to determine the PID of the
service process when waiting for the new background service to start
to ensure that the new service is running.

Log: fix front-end startup stuck
Bug: https://pms.uniontech.com/bug-view-239747.html